Version 1.8.0: Repository cleanup and Python package fixes
- Bump version from 1.7.0 to 1.8.0
- Remove redundant files and directories (~50MB saved)
- Fix Python package type conversion issues
- Update documentation to use uv pip install . instead of maturin develop
- All tests passing (44/44)
- CLI and Python package working correctly

+++ tidy-viewer-py/src/formatting.rs
@@ -6,7 +6,7 @@ use owo_colors::OwoColorize;
 use unicode_width::UnicodeWidthStr;
 
 use crate::types::{FormatOptions, ColorScheme};
-use crate::datatype::{format_strings, calculate_column_width, is_negative_number, is_na};
+use tidy_viewer_core::{format_strings, calculate_column_width, is_negative_number, is_na};
 
 /// Main entry point for formatting tabular data
 pub fn format_table(
@@ -69,7 +69,7 @@ pub fn format_table(
                 col,
                 options.min_col_width,
                 options.max_col_width,
-                options.significant_figures as u8,
+                options.significant_figures as i64,
                 options.preserve_scientific,
                 options.max_decimal_width,
             )
@@ -110,7 +110,7 @@ pub fn format_table(
                     col,
                     options.min_col_width,
                     options.max_col_width,
-                    options.significant_figures as u8,
+                    options.significant_figures as i64,
                     options.preserve_scientific,
                     options.max_decimal_width,
                 )
